Added more comments and moved things around for simplicity.
